requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
selenium==4.24.0
pillow==10.4.0
transformers==4.44.2
torch==2.4.1
pandas==2.2.2
matplotlib==3.9.2
seaborn==0.13.2
opencv-python==4.10.0.84